+++ CHANGES.md
@@ -5,6 +5,7 @@
 ### titiler.core
 
 * fix `TerrainRGB` algorithm name (author @JinIgarashi, https://github.com/developmentseed/titiler/pull/804)
+* add more tests for `RescalingParams` and `HistogramParams` dependencies
 
 ## 0.18.0 (2024-03-22)